backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.11.2
requests==2.31.0
python-dotenv==1.0.0